projects/requirements.txt
streamlit
opencv-python
torch
torchvision
pillow
